requirements.txt
pandas==2.1.4
requests==2.31.0
dbt-core==1.7.9
dbt-postgres==1.7.9
apache-airflow-providers-amazon==8.13.0
psycopg2-binary==2.9.9